refactor theme toggler into shared component
